[Security] WIP refactoring of the security chapter
This makes the security chapter cover much more core functionality, and written a bit more like a tutorial.

This moves much into cookbooks as well.

This is a WIP as there are still a few things on the previous version of the docs that aren't included here. I also have a list of things that are missing from both (and are pretty important).